Reorganize backend
