--- backend_py/build.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-mkdir build -p
-cc -fPIC -shared -o ${PWD}/build/camera_helper.so ${PWD}/src/camera_helper.c
