+++ backend_py/src/__init__.py
@@ -6,18 +6,9 @@ from flask import Flask, jsonify, request
 from flask_cors import CORS
 from gevent.pywsgi import WSGIServer
 
-from .enumeration import *
-from .camera_helper_loader import *
-from .schemas import *
-from .device import *
-from .stream import *
-from .settings import SettingsManager
-from .broadcast_server import BroadcastServer
-from .device_manager import DeviceManager
-from .lights.light_manager import LightManager
-from .lights.utils import create_pwm_controllers
-from .preferences.preferences_manager import PreferencesManager
-from .preferences.preference_schemas import SavedPrefrencesSchema, SavedPrefrences
+from .websockets.broadcast_server import BroadcastServer
+
+from .services import *
 
 import logging
