+++ backend_py/src/services/__init__.py
@@ -1,0 +1,3 @@
+from .cameras import *
+from .lights import *
+from .preferences import *
