+++ backend_py/src/services/cameras/camera_helper/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+mkdir build -p
+cc -fPIC -shared -o ${PWD}/build/camera_helper.so ${PWD}/camera_helper.c
