+++ backend_py/src/services/cameras/camera_helper/camera_helper.c
