+++ backend_py/src/services/cameras/device_utils.py
