+++ backend_py/src/services/cameras/ehd.py
@@ -1,8 +1,8 @@
 from typing import Dict
-from ..enumeration import DeviceInfo
-from ..device import Device, Option, ControlTypeEnum
-from ..camera_types import H264Mode
-from .. import ehd_controls as xu
+from .enumeration import DeviceInfo
+from .device import Device, Option, ControlTypeEnum
+from .camera_types import H264Mode
+from . import ehd_controls as xu
 
 class EHDDevice(Device):
     '''
