+++ backend_py/src/services/cameras/ehd_controls.py
