+++ backend_py/src/services/cameras/enumeration.py
