+++ backend_py/src/services/cameras/saved_types.py
