+++ backend_py/src/services/cameras/schemas.py
