+++ backend_py/src/services/cameras/stream.py
