+++ backend_py/src/services/cameras/stream_utils.py
