+++ backend_py/src/services/lights/light_manager.py
