+++ backend_py/src/services/lights/light_types.py
