+++ backend_py/src/services/lights/schemas.py
