+++ install_requirements.sh
@@ -3,8 +3,5 @@
 # install python
 sudo apt-get install python3 python3-venv -y
 
-# install golang
-sudo apt-get install golang -y
-
 # install gstreamer
 sudo apt-get install -y libglib2.0-dev libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev gstreamer1.0-tools gstreamer1.0-x gstreamer1.0-plugins-base gstreamer1.0-plugins-good gstreamer1.0-plugins-bad gstreamer1.0-libav
