--- system_api/README.md
@@ -1,13 +1,0 @@
-go build && clear && ./system_api
-
-## Usage
-
-#### Available Commands
-
-Here is a list of all the available commands:
-
-| Script      | Description                                                                                   |
-| ----------- | --------------------------------------------------------------------------------------------- |
-| go mod tidy | Installs all the dependencies listed in `go.mod`.                                             |
-| go run .    | Runs the application without building it.                                                     |
-| go build .  | Builds the application into an executable file for different platforms using the Go compiler. |
