--- system_api/go.mod
@@ -1,22 +1,0 @@
-module system_api
-
-go 1.18
-
-require (
-	github.com/gorilla/handlers v1.5.1
-	github.com/gorilla/mux v1.8.0
-	github.com/shirou/gopsutil v3.21.11+incompatible
-	github.com/sirupsen/logrus v1.9.3
-	github.com/theojulienne/go-wireless v1.2.0
-	gopkg.in/natefinch/lumberjack.v2 v2.2.1
-)
-
-require (
-	github.com/felixge/httpsnoop v1.0.1 // indirect
-	github.com/go-ole/go-ole v1.2.6 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/tklauser/go-sysconf v0.3.12 // indirect
-	github.com/tklauser/numcpus v0.6.1 // indirect
-	github.com/yusufpapurcu/wmi v1.2.3 // indirect
-	golang.org/x/sys v0.11.0 // indirect
-)
